Refactore auth store
